move faq entities
