+++ src/main/webapp/app/communication/entities/faq-category.model.ts
