+++ src/main/webapp/app/communication/entities/faq.model.ts
