+++ src/test/javascript/spec/component/shared/monaco-editor/monaco-editor-communication-action.integration.spec.ts
@@ -29,7 +29,7 @@ import { ReferenceType } from 'app/communication/metis.util';
 import { Attachment } from 'app/entities/attachment.model';
 import dayjs from 'dayjs/esm';
 import { FaqReferenceAction } from 'app/shared/monaco-editor/model/actions/communication/faq-reference.action';
-import { Faq } from 'app/entities/faq.model';
+import { Faq } from 'app/communication/entities/faq.model';
 import { FileService } from 'app/shared/http/file.service';
 import { MockFileService } from '../../../helpers/mocks/service/mock-file.service';
